requirements.txt
blessed==1.14.2
botocore==1.10.2
cement==2.8.2
certifi==2018.1.18
chardet==3.0.4
colorama==0.3.7
defusedxml==0.5.0
Django==2.0.3
django-braces==1.12.0
django-oauth-toolkit==1.0.0
djangorestframework==3.7.7
djangorestframework-gis==0.12
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
dry-rest-permissions==0.1.10
idna==2.6
jmespath==0.9.3
oauthlib==2.0.7
pathspec==0.5.5
pbr==3.1.1
psycopg2==2.7.4
PyJWT==1.6.1
python-dateutil==2.6.1
python3-openid==3.1.0
pytz==2018.3
PyYAML==3.12
requests==2.13.0
requests-oauthlib==0.8.0
semantic-version==2.5.0
six==1.11.0
stevedore==1.27.1
tabulate==0.7.5
termcolor==1.1.0
urllib3==1.22
uWSGI==2.0.17
virtualenv==15.1.0
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.2
wcwidth==0.1.7
websocket-client==0.47.0
